AG-12563 - Fix circular dependencies
