+++ packages/ag-charts-enterprise/src/features/annotations/cross-line/crossLineProperties.ts
@@ -2,7 +2,7 @@ import { type PixelSize, _ModuleSupport, type _Scene } from 'ag-charts-community
 
 import { Annotation, AxisLabel, Handle, LineStyle, LineTextProperties, Stroke, Value } from '../annotationProperties';
 import { type AnnotationContext, type AnnotationOptionsColorPickerType, AnnotationType } from '../annotationTypes';
-import { getComputedLineDash, getLineStyle } from '../utils/styles';
+import { getLineCap, getLineDash } from '../properties/util';
 import { validateDatumValue } from '../utils/validation';
 
 const { OBJECT, STRING, BaseProperties, Validate, isObject } = _ModuleSupport;
@@ -36,13 +36,11 @@ export class HorizontalLineProperties extends Annotation(Value(Handle(AxisLabel(
     }
 
     getLineDash(): PixelSize[] | undefined {
-        const styleType = getLineStyle(this.lineDash, this.lineStyle);
-        return this.lineDash ?? this.computedLineDash ?? getComputedLineDash(this.strokeWidth ?? 1, styleType);
+        return getLineDash(this.lineDash, this.computedLineDash, this.lineStyle, this.strokeWidth);
     }
 
     getLineCap(): _Scene.ShapeLineCap | undefined {
-        const styleType = getLineStyle(this.lineDash, this.lineStyle);
-        return this.lineCap ?? styleType === 'dotted' ? 'round' : undefined;
+        return getLineCap(this.lineCap, this.lineDash, this.lineStyle);
     }
 }
 
@@ -75,13 +73,11 @@ export class VerticalLineProperties extends Annotation(Value(Handle(AxisLabel(St
     }
 
     getLineDash(): PixelSize[] | undefined {
-        const styleType = getLineStyle(this.lineDash, this.lineStyle);
-        return this.lineDash ?? this.computedLineDash ?? getComputedLineDash(this.strokeWidth ?? 1, styleType);
+        return getLineDash(this.lineDash, this.computedLineDash, this.lineStyle, this.strokeWidth);
     }
 
     getLineCap(): _Scene.ShapeLineCap | undefined {
-        const styleType = getLineStyle(this.lineDash, this.lineStyle);
-        return this.lineCap ?? styleType === 'dotted' ? 'round' : undefined;
+        return getLineCap(this.lineCap, this.lineDash, this.lineStyle);
     }
 }
